Refactor route structure by moving routes to index file and removing old routes file
